перенес из conftest.py в helpers.py три метода для генерации и создания пользователя
