+++ conftest.py
@@ -1,23 +1,8 @@
 import pytest
 import requests
-import random
-import string
 from data import BASE_URL, REGISTER_URL, LOGIN_URL
-from auth import get_auth_headers, get_tokens #добавил
-
-
-# # Фикстура для получения базового URL
-# @pytest.fixture(scope="session")
-# def base_url():
-#     return BASE_URL
-
-# Функция для генерации случайного email
-def random_email():
-    return f"{''.join(random.choices(string.ascii_lowercase + string.digits, k=10))}@example.com"
-
-# Функция для генерации случайного имени пользователя
-def random_name():
-    return ''.join(random.choices(string.ascii_letters, k=10))
+from auth import get_auth_headers, get_tokens
+from helpers import random_email, random_name, ensure_user_exists
 
 # Фикстура для создания уникального пользователя
 @pytest.fixture(scope="function")
@@ -29,12 +14,6 @@ def unique_user():
     }
     return user
 
-def ensure_user_exists(user):
-    try:
-        get_tokens(user["email"], user["password"])
-    except requests.exceptions.HTTPError:
-        response = requests.post(REGISTER_URL, json=user)
-        assert response.status_code == 200, f"Failed to register user: {response.status_code} {response.text}"
 
 # Фикстура для использования зарегистрированного пользователя
 @pytest.fixture(scope="function")
